Refactor settings and voting components to use react-query for data fetching and mutation handling, replacing trpc client usage. Update API key handling in the backend and clean up unused redis publish calls in poll router.
