+++ src/components/SettingsFormWrapper.tsx
@@ -1,29 +1,14 @@
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { httpBatchLink } from '@trpc/client';
-import { useState } from 'react';
-import { trpcReact } from '@/lib/trpc/client';
+import { QueryClientProvider } from '@tanstack/react-query';
 import { SettingsForm } from './settingsForm';
-import { Toaster } from './ui/toaster';
+import { queryClient } from '@/lib/trpcs';
+import { Toaster } from './ui/sonner';
 
 const VoteFormWrapper = () => {
-  const [queryClient] = useState(() => new QueryClient());
-  const [trpcClient] = useState(() =>
-    trpcReact.createClient({
-      links: [
-        httpBatchLink({
-          url: '/api/trpc'
-        })
-      ]
-    })
-  );
-
   return (
-    <trpcReact.Provider client={trpcClient} queryClient={queryClient}>
       <QueryClientProvider client={queryClient}>
         <SettingsForm></SettingsForm>
         <Toaster />
       </QueryClientProvider>
-    </trpcReact.Provider>
   );
 };
 export default VoteFormWrapper;
