+++ src/pages/settings.astro
@@ -1,6 +1,4 @@
 ---
-import '@/styles/globals.css';
-
 import SiteLayout from '@/layouts/CardLayout.astro';
 import SettingsFormWrapper from '@/components/SettingsFormWrapper';
 if (!Astro.locals.session) return Astro.redirect('/api/auth/signin');
